feat: [SPLT-1549] Realizado ajustes de revisao do pr

+++ trip-schema/src/main/java/global/maplink/trip/schema/solution/TripResponse.java
@@ -24,7 +24,7 @@ public class TripResponse {
     private final String clientId;
     private final Long totalDistance;
     private final Long totalNominalDuration;
-    private final Double averageSpeed;
+    private final BigDecimal averageSpeed;
     private final BigDecimal tollCosts;
     private final BigDecimal routeFreightCost;
     private final List<SolutionLeg> legs;
